+++ R/spread2.R
@@ -327,14 +327,12 @@ if (getRversion() >= "3.1.0") {
 #'
 #' @author Eliot McIntire and Steve Cumming
 #' @export
-#' @importFrom bit bit
 #' @importFrom checkmate assert assertClass assertNumeric
 #' @importFrom checkmate checkClass checkDataTable checkLogical checkNumeric checkScalarNA
 #' @importFrom checkmate qassert
 #' @importFrom data.table := alloc.col as.data.table copy data.table is.data.table
 #' @importFrom data.table rbindlist set setattr setcolorder setkeyv setnames uniqueN
 #' @importFrom fastmatch fmatch
-#' @importFrom ff ff
 #' @importFrom fpCompare %<=% %>>%
 #' @importFrom magrittr %>%
 #' @importFrom quickPlot Plot
@@ -464,11 +462,7 @@ spread2 <- function(landscape, start = ncell(landscape) / 2 - ncol(landscape) / 
   if (!is.data.table(start)) {
     # A "new" entry into spread2 -- need to set up stuff
     if (canUseAvailable) {
-      if (smallRaster) {
-        notAvailable <- bit(ncells)
-      } else {
-        notAvailable <- ff(vmode = "boolean", FALSE, length = ncells)
-      }
+      notAvailable <- logical(ncells)
       notAvailable[start] <- TRUE
     }