+++ R/spread3.R
@@ -59,7 +59,7 @@ spread3 <- function(start, rasQuality, rasAbundance, advectionDir,
                     advectionMag, meanDist, plot.it = 2,
                     minNumAgents = 50, verbose = getOption("LandR.verbose", 0),
                     saveStack = NULL) {
-
+  dtThr <- data.table::getDTthreads()
   testEquivalentMetadata(rasAbundance, rasQuality)
 
   if (is(advectionDir, "Raster")) {
@@ -105,6 +105,8 @@ spread3 <- function(start, rasQuality, rasAbundance, advectionDir,
   rasIterations[start$pixels] <- 0
 
   while (abundanceDispersing > minNumAgents) {
+
+    if (dtThr == 1 && data.table::getDTthreads() != 1) data.table::setDTthreads(1)
     b <- spread2(landscape = rasQuality, start = start,
                  spreadProb = 1, iterations = 1, asRaster = FALSE,
                  returnDistances = TRUE, returnFrom = TRUE,
@@ -301,19 +303,8 @@ testEquivalentMetadata <- function(...) {
 }
 
 #' @export
+#' @importFrom raster compareRaster
 testEquivalentMetadata.Raster <- function(...) {
-  d <- list(...)
-  res <- lapply(d[-1], function(x) {
-    if (!identical(d[[1]]@extent, x@extent)) {
-      stop("rasAbundance and rasQuality must have same extent")
-    }
-    if (!identical(d[[1]]@crs, x@crs)) {
-      stop("rasAbundance and rasQuality must have same crs")
-    }
-    if (!identical(res(d[[1]]), res(x))) {
-      stop("rasAbundance and rasQuality must have same res")
-    }
-
-  })
+  compareRaster(..., orig = TRUE)
   return(invisible())
 }