+++ NEWS.md
@@ -1,5 +1,21 @@
 Known issues: https://github.com/PredictiveEcology/SpaDES.tools/issues
 
+version 0.3.5
+=============
+
+We have identified a non-trivial bug in `spread` when using `allowOverlap = TRUE`.
+Until we can patch this bug, using this argument value will generate an error.
+We are currently working on a fix, which may take some time.
+
+## Dependency changes
+* removed orphaned package `bit`; its reverse dependencies, `ff` and `ffbase`, moved to Suggests
+
+## Bugfixes
+* none
+
+## Enhancements
+* New values for `returnIndices` arg in `spread`, allowing for `numeric`. New option, `2` allows for the fastest return possible, only `pixelIndices`
+
 version 0.3.4
 =============